+++ .project
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
 	<name>gradle-demo</name>
-	<comment></comment>
+	<comment>Project gradle-demo created by Buildship.</comment>
 	<projects>
 	</projects>
 	<buildSpec>
