+++ build.gradle
@@ -15,7 +15,7 @@ buildscript{
 apply plugin : 'application'
 apply plugin : 'eclipse'
 
-mainClassName = 'com.sanjeev.gradledemo.main.GradleDemoMain'
+mainClassName = 'sanrockzz.gradledemo.main.GradleDemoMain'
 version = '1.0'
 
 repositories {
@@ -31,6 +31,29 @@ jar {
                    'Implementation-Version': version
     }
 }
+
+dependencies {
+
+	compile "org.springframework.boot:spring-boot-starter-web:1.5.8.RELEASE"
+   	compile "org.json:json:20171018"
+	   	
+   	compile "com.fasterxml.jackson.core:jackson-databind:2.9.3"
+	compile "com.fasterxml.jackson.core:jackson-annotations:2.9.3"
+	compile "com.fasterxml.jackson.core:jackson-core:2.9.3"
+	
+   	compile "io.springfox:springfox-swagger2:2.9.2"
+	compile "io.springfox:springfox-swagger-ui:2.9.2"
+	compile "org.webjars:jquery:2.2.4"
+	
+	compile "org.springframework.ws:spring-ws-core:2.4.2.RELEASE"
+	compile "org.springframework:spring-oxm:3.0.4.RELEASE"
+		   	
+   	testCompile "junit:junit:4.12"
+   	testCompile "org.jmockit:jmockit:1.9"
+   	testCompile"org.springframework.boot:spring-boot-starter-test:1.5.8.RELEASE"
+	
+	}
+
 
 task copyToLib(type: Copy) {
     	from configurations.runtime
@@ -48,28 +71,3 @@ task copyDockerFile(type: Copy) {
 
 build.dependsOn(copyToLib, copyDockerFile, copyJar)
 
-dependencies {
-
-	compile("org.springframework.boot:spring-boot-starter-web:1.5.8.RELEASE")
-
-   	compile("org.json:json:20171018")
-	   	
-   	compile group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.9.3'
-	compile group: 'com.fasterxml.jackson.core', name: 'jackson-annotations', version: '2.9.3'
-	compile group: 'com.fasterxml.jackson.core', name: 'jackson-core', version: '2.9.3'
-	
-	
-   	compile group: 'io.springfox', name: 'springfox-swagger2', version: '2.0.1'
-	compile group: 'io.springfox', name: 'springfox-swagger-ui', version: '2.0.1'
-
-	compile group: 'org.webjars', name: 'jquery', version: '2.2.4'
-	
-	compile group: 'org.springframework.ws', name: 'spring-ws-core', version: '2.4.2.RELEASE'
-	
-	compile group: 'org.springframework', name: 'spring-oxm', version: '3.0.4.RELEASE'
-		   	
-   	testCompile ("junit:junit:4.12")
-   	testCompile ("org.jmockit:jmockit:1.9")
-   	testCompile("org.springframework.boot:spring-boot-starter-test:1.5.8.RELEASE")
-	
-	}
