--- src/main/java/com/sanjeev/gradledemo/config/AppConfig.java
@@ -1,38 +1,0 @@
-package com.sanjeev.gradledemo.config;
-
-import org.springframework.boot.SpringBootConfiguration;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.oxm.jaxb.Jaxb2Marshaller;
-import org.springframework.ws.client.core.WebServiceTemplate;
-
-/**
- *
- * Source AppConfig.java created on Sep 9, 2018
- *
- * @author : Sanjeev Saxena
- * @email : sanrocks123@gmail.com
- * @version : 1.0
- */
-
-@ComponentScan("com.sanjeev")
-@EnableAutoConfiguration
-@SpringBootConfiguration
-public class AppConfig {
-
-    @Bean
-    Jaxb2Marshaller jaxb2Marshaller() {
-        final Jaxb2Marshaller jaxb2Marshaller = new Jaxb2Marshaller();
-        jaxb2Marshaller.setContextPath("com.ms.nao.csid.model");
-        return jaxb2Marshaller;
-    }
-
-    @Bean
-    public WebServiceTemplate webServiceTemplate() {
-        final WebServiceTemplate webServiceTemplate = new WebServiceTemplate();
-        webServiceTemplate.setMarshaller(jaxb2Marshaller());
-        webServiceTemplate.setUnmarshaller(jaxb2Marshaller());
-        return webServiceTemplate;
-    }
-}
