+++ src/main/java/sanrockzz/creditsuisse/assignment/ItemRepository.java
@@ -2,7 +2,7 @@
  * Copyright (c) @Sanjeev Saxena 2017. All Rights Reserved.
  */
 
-package creditsuisse.assignment;
+package sanrockzz.creditsuisse.assignment;
 
 /**
  * Java Source ItemRepository.java created on Nov 19, 2019
