+++ src/main/java/sanrockzz/gradledemo/config/AppConfig.java
@@ -1,0 +1,17 @@
+package sanrockzz.gradledemo.config;
+
+import org.springframework.context.annotation.Configuration;
+
+/**
+ *
+ * Source AppConfig.java created on Sep 9, 2018
+ *
+ * @author : Sanjeev Saxena
+ * @email : sanrocks123@gmail.com
+ * @version : 1.0
+ */
+
+@Configuration
+public class AppConfig {
+
+}
