+++ src/main/java/sanrockzz/gradledemo/dto/FileDto.java
@@ -2,7 +2,7 @@
  * Copyright (c) @Sanjeev Saxena 2017. All Rights Reserved.
  */
 
-package com.sanjeev.gradledemo.dto;
+package sanrockzz.gradledemo.dto;
 
 import org.springframework.web.multipart.MultipartFile;
