+++ src/main/java/sanrockzz/gradledemo/main/GradleDemoMain.java
@@ -2,10 +2,12 @@
  * Copyright (c) @Sanjeev Saxena 2017. All Rights Reserved.
  */
 
-package com.sanjeev.gradledemo.main;
+package sanrockzz.gradledemo.main;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
 /**
  * Main entry point
@@ -16,7 +18,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
  *
  */
 
-@SpringBootApplication
+@EnableSwagger2
+@SpringBootApplication(scanBasePackages = "sanrockzz")
 public class GradleDemoMain {
 
     /**
@@ -24,6 +27,7 @@ public class GradleDemoMain {
      */
     public static void main(String[] args) {
         final SpringApplication app = new SpringApplication(GradleDemoMain.class);
+        // app.setBanner(new ResourceBanner(new PathResource()));
         app.run(args);
     }
