+++ src/main/resources/application.properties
@@ -3,5 +3,6 @@
 # Author		: Sanjeev Saxena
 #
 
-
-server.port=1000
+server.port						= 1000
+server.servlet.context-path		= /tagging-service
+spring.banner.location			= classpath:/banner.txt
