+++ src/main/resources/banner.txt
@@ -1,0 +1,9 @@
+
+___________                     .__                   _________                  .__              
+\__    ___/____     ____   ____ |__| ____    ____    /   _____/ ______________  _|__| ____  ____  
+  |    |  \__  \   / ___\ / ___\|  |/    \  / ___\   \_____  \_/ __ \_  __ \  \/ /  |/ ___\/ __ \ 
+  |    |   / __ \_/ /_/  > /_/  >  |   |  \/ /_/  >  /        \  ___/|  | \/\   /|  \  \__\  ___/ 
+  |____|  (____  /\___  /\___  /|__|___|  /\___  /  /_______  /\___  >__|    \_/ |__|\___  >___  >
+               \//_____//_____/         \//_____/           \/     \/                    \/    \/ 
+
+
