+++ src/test/java/creditsuisse/ShoppingCartTest.java
@@ -7,9 +7,9 @@ package creditsuisse;
 import org.junit.Assert;
 import org.junit.Test;
 
-import creditsuisse.assignment.ItemTestDataRepositoryImpl;
-import creditsuisse.assignment.ShoppingOperation;
-import creditsuisse.assignment.ShoppingOperationsImpl;
+import sanrockzz.creditsuisse.assignment.ItemTestDataRepositoryImpl;
+import sanrockzz.creditsuisse.assignment.ShoppingOperation;
+import sanrockzz.creditsuisse.assignment.ShoppingOperationsImpl;
 
 /**
  * Java Source OperationTest.java created on Nov 19, 2019
